requirements.txt
web3
flask
flask_cors
kafka-python
